factor out the message blurb
